pyproject.toml
[project]
name = "discord-llm-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mypy>=1.11",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typer>=0.21.0",
    "types-requests>=2.32.4.20250913",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
indent-style = "tab"

[tool.ruff.lint]
extend-select = ["I"]
